Add files via upload
Payment 100% done

+++ src/Cash.java
@@ -3,7 +3,7 @@
 public class Cash implements Payment{
 
     @Override
-    public boolean Pay(int amount) {
+    public boolean Pay(Double amount) {
         System.out.println("Payment Done Using Cash On Delivery");
         return true;
     }
