+++ src/Credit.java
@@ -4,7 +4,7 @@
 public class Credit implements Payment{
 
     @Override
-    public boolean Pay(int amount) {
+    public boolean Pay(Double amount) {
         System.out.println("Payment Done Using Credit Card");
         return true;
     }
