+++ src/Wallet.java
@@ -3,7 +3,7 @@
 public class Wallet implements Payment{
 
     @Override
-    public boolean Pay(int amount) {
+    public boolean Pay(Double amount) {
         System.out.println("Payment Done Using Cash from Wallet");
         return true;
     }
